[conversion] move to core/ folder and expose /api/conversion
