+++ sysadmin/configs/nginx/default-docker
@@ -87,6 +87,9 @@ server {
 	location /api/dispatcher/ {
 		proxy_pass  http://$host:8080;
 	}
+	location /api/conversion/ {
+		proxy_pass  http://$host:8080;
+	}
 
 	# shared_folder, run static data
 	location /api/core/shared_folder/run/ {
