+++ sysadmin/configs/nginx/default-prod
@@ -119,6 +119,9 @@ server {
     location /api/dispatcher/ {
         proxy_pass  http://$host:8080;
     }
+    location /api/conversion/ {
+        proxy_pass  http://$host:8080;
+    }
 
     # shared_folder, run static data
     location /api/core/shared_folder/run/ {
